+++ src/pine/ast/order.clj
@@ -5,7 +5,7 @@
         current (state :current)
         columns (map #(-> %1
                           (assoc :alias (or (:alias %1) current))
-                          (assoc :index i))
+                          (assoc :operation-index i))
                      value)]
     (-> state
         (update :order into columns))))